+++ src/components/explorer-link/explorer-link.tsx
@@ -1,6 +1,7 @@
-import { Component, Prop, h } from '@stencil/core';
-import { getIconHtmlFromIconDefinition } from 'utils/icons/getIconHtmlFromIconDefinition';
-import { IconDefinition, faArrowUpRightFromSquare } from '@fortawesome/free-solid-svg-icons';
+import type { IconDefinition} from '@fortawesome/free-solid-svg-icons';
+import { faArrowUpRightFromSquare } from '@fortawesome/free-solid-svg-icons';
+import { Component, Element, h, Prop, State } from '@stencil/core';
+import classNames from 'classnames';
 
 @Component({
   tag: 'explorer-link',
@@ -8,18 +9,23 @@ import { IconDefinition, faArrowUpRightFromSquare } from '@fortawesome/free-soli
   shadow: true,
 })
 export class ExplorerLink {
-  @Prop() class: string = 'explorer-link';
+  @Prop() class: string;
   @Prop() dataTestId?: string;
   @Prop() icon?: IconDefinition;
   @Prop() link: string;
   @Prop() text?: string;
 
-  render() {
-    const icon = getIconHtmlFromIconDefinition(this.icon ?? faArrowUpRightFromSquare);
+  @Element() hostElement: HTMLElement;
+  @State() hasSlotContent: boolean = false;
 
+  componentWillLoad() {
+    this.hasSlotContent = !!this.hostElement.querySelector('[slot="content"]');
+  }
+
+  render() {
     return (
-      <a data-testid={this.dataTestId} href={this.link} target="_blank" class={this.class} rel="noreferrer">
-        {this.text ?? <span innerHTML={icon}></span>}
+      <a data-testid={this.dataTestId} href={this.link} target="_blank" class={classNames(this.class, 'explorer-link')} rel="noreferrer">
+        {this.hasSlotContent ? <slot name="content"></slot> : this.text ?? <fa-icon icon={this.icon ?? faArrowUpRightFromSquare}></fa-icon>}
       </a>
     );
   }