+++ src/components/toasts-list/components/transaction-toast/components/transaction-toast-details/transaction-toast-details.tsx
@@ -1,5 +1,7 @@
-import { Component, Fragment, JSX, Prop, h } from '@stencil/core';
-import { ITransaction } from '../../transaction-toast.type';
+import type { JSX } from '@stencil/core';
+import { Component, Fragment, h, Prop } from '@stencil/core';
+
+import type { ITransaction } from '../../transaction-toast.type';
 
 @Component({
   tag: 'transaction-toast-details',
@@ -20,8 +22,8 @@ export class TransactionDetails {
       <Fragment>
         <div class="status-title">{this.processedTransactionsStatus}</div>
 
-        {this.transactions.map(({ hash, status }) => (
-          <transaction-toast-details-body transactionClass={this.transactionClass} hash={hash} status={status} key={hash} />
+        {this.transactions.map(({ hash, status, link }) => (
+          <transaction-toast-details-body transactionClass={this.transactionClass} hash={hash} status={status} link={link} key={hash} />
         ))}
       </Fragment>
     );