+++ src/components/toasts-list/components/transaction-toast/transaction-toast.tsx
@@ -1,5 +1,7 @@
-import { Component, Prop, Event, h, EventEmitter, JSX } from '@stencil/core';
-import { ITransactionProgressState, IToastDataState, ITransaction } from './transaction-toast.type';
+import type { EventEmitter, JSX } from '@stencil/core';
+import { Component, Event, h,Prop } from '@stencil/core';
+
+import type { IToastDataState, ITransaction,ITransactionProgressState } from './transaction-toast.type';
 
 @Component({
   tag: 'transaction-toast',
@@ -20,11 +22,7 @@ export class TransactionToast {
   render() {
     return (
       <transaction-toast-wrapper wrapperId={`toast-${this.toastId}`} wrapperClass={this.wrapperClass}>
-        <transaction-toast-progress
-          key={this.toastId}
-          progressClass={this.transactionProgressState?.progressClass}
-          currentRemaining={this.transactionProgressState?.currentRemaining}
-        >
+        <transaction-toast-progress key={this.toastId} startTime={this.transactionProgressState?.startTime} endTime={this.transactionProgressState?.endTime}>
           <transaction-toast-content
             onDeleteToast={() => this.handleDeleteToast.emit(this.toastId)}
             processedTransactionsStatus={this.processedTransactionsStatus}