+++ src/components/toasts-list/components/transaction-toast/transaction-toast.type.ts
@@ -1,6 +1,6 @@
 // match these interfaces with src/path-to-file.type.ts from sdk-dapp
-import { IconName } from '@fortawesome/free-solid-svg-icons';
-import { JSX } from '@stencil/core';
+import type { IconName } from '@fortawesome/free-solid-svg-icons';
+import type { JSX } from '@stencil/core';
 
 export interface IToastDataState {
   id: string;
@@ -12,12 +12,15 @@ export interface IToastDataState {
 
 export interface ITransactionProgressState {
   progressClass?: string;
-  currentRemaining: number;
+  startTime: number;
+  endTime: number;
+  isCrossShard: boolean;
 }
 
 export interface ITransaction {
   hash: string;
   status: string;
+  link: string;
 }
 
 export interface ITransactionToast {