+++ src/components/transactions-table/components/transaction-account/components/transaction-account-name/transaction-account-name.tsx
@@ -1,4 +1,4 @@
-import { Component, Prop, h } from '@stencil/core';
+import { Component, h, Prop } from '@stencil/core';
 import classNames from 'classnames';
 
 @Component({
@@ -7,20 +7,21 @@ import classNames from 'classnames';
   shadow: true,
 })
 export class TransactionAccountName {
-  @Prop() class?: string = 'transaction-account-name';
+  @Prop() address: string;
+  @Prop() class?: string;
   @Prop() dataTestId?: string;
   @Prop() description: string;
-  @Prop() name: string;
+  @Prop() name?: string;
 
   render() {
     if (this.name) {
       return (
-        <span class={classNames('text-truncate', this.class)} data-testid={this.dataTestId} title={this.description}>
+        <div class={classNames('text-truncate', this.class, 'transaction-account-name')} data-testid={this.dataTestId} title={this.description}>
           {this.name}
-        </span>
+        </div>
       );
     }
 
-    return <span class="trim">{this.description}</span>;
+    return <trim-text text={this.address} class={classNames(this.class, 'transaction-account-name')} dataTestId={this.dataTestId}></trim-text>;
   }
 }