+++ src/components/transactions-table/components/transaction-method/transaction-method.tsx
@@ -1,6 +1,6 @@
 import { Component, h, Prop } from '@stencil/core';
-import { DataTestIdsEnum } from 'constants/dataTestIds.enum';
 import classNames from 'classnames';
+import { DataTestIdsEnum } from 'constants/dataTestIds.enum';
 
 @Component({
   tag: 'transaction-method',
@@ -8,15 +8,19 @@ import classNames from 'classnames';
   shadow: true,
 })
 export class TransactionMethod {
-  @Prop() class?: string = 'transaction-method';
+  @Prop() class?: string;
   @Prop() actionDescription: string;
   @Prop() method: string;
 
   render() {
     return (
-      <span class={classNames('badge badge-secondary badge-pill font-weight-light p-0', this.class)} data-testid={DataTestIdsEnum.method} title={this.actionDescription}>
+      <span
+        class={classNames('badge badge-secondary badge-pill font-weight-light transaction-method', this.class)}
+        data-testid={DataTestIdsEnum.method}
+        title={this.actionDescription}
+      >
         <div class="badge badge-secondary badge-pill">
-          <div class="text-truncate text-capitalize text-white p-1">{this.method}</div>
+          <div class="text-truncate text-capitalize text-white">{this.method}</div>
         </div>
       </span>
     );